pyproject.toml
[project]
name = "line-deployment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "python-multipart>=0.0.9",
    "firebase-admin>=7.1.0",
    "google-cloud-storage>=3.8.0",
    "line-bot-sdk>=3.22.0",
    "openai>=2.16.0",
    "opencv-python-headless>=4.13.0.90",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "ultralytics>=8.4.8",
    "uvicorn>=0.40.0",
]
